BOL-375 refactor: group 관련 model(in data module) package 분리
