+++ data/src/main/java/com/yapp/bol/data/model/group/CheckGroupJoinByAccessCodeResponse.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.group.response
+package com.yapp.bol.data.model.group
 
 import com.google.gson.annotations.SerializedName
