+++ data/src/main/java/com/yapp/bol/data/model/group/GroupSearchApiResponse.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.group.response
+package com.yapp.bol.data.model.group
 
 data class GroupSearchApiResponse(
     val content: List<GroupListItem>,
