+++ data/src/main/java/com/yapp/bol/data/model/group/RandomImageResponse.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.group.response
+package com.yapp.bol.data.model.group
 
 data class RandomImageResponse(
     val url: String
