+++ data/src/main/java/com/yapp/bol/data/model/group/UserRankApiResponse.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.rank
+package com.yapp.bol.data.model.group
 
 import com.google.gson.annotations.SerializedName
